Add a Dockerfile
This is a similar Dockerfile to the one I made for
https://github.com/akkornel/livia_helloworld.

It uses the Rocker project's R 4.3.2 image as a base, which itself uses
Ubuntu 22.04 LTS as a base.  On top of that, we install the Microsoft
ODBC Driver 18 (from their package repo), as well as the ODBC library
needed for R's ODBC package.

We also do an `apt-get upgrade`, just in case.

One small hack: The app uses the ODBC driver named "SQL Server".  We
change this (using `sed`) to the name of the driver provided by
Microsoft.

Afterwards, we install the R `renv` package, and then restore from the
Renv package lock file.  The container's command runs Shiny, exposing it
on port 3838.

We have a couple of environment variables used by the container, for
which we set some defaults.