Downgrade container ODBC driver from 18 to 17
This change switches the container from using the Microsoft ODBC Driver
18 for SQL server, down to the Microsoft ODBC Driver 17 for SQL Server.
The package name changes, and the driver name changes.

This change is being made because Driver 18 introduced a breaking
change: `Encytpion=True` is set by default, unless overridden.

Attempting to connect with Driver 18 gives this error:

```
[Microsoft][ODBC Driver 18 for SQL Server]SSL Provider: [error:0A000086:SSL routines::certificate verify failed:EE certificate key too weak]
```

See https://techcommunity.microsoft.com/t5/sql-server-blog/odbc-driver-18-0-for-sql-server-released/ba-p/3169228
for more information on the change.

In order to move to Driver 18, the SQL Server needs a configuration
change.