Actions: Update actions & switch to commit IDs
This updates a number of our GitHub Actions, and switches from using tag
names to using commit IDs.

Actions are being upgraded as follows:

* actions/checkout: v4 → v4.2.2
* docker/setup-buildx-action: v3.1.0 → v3.10.0
* docker/login-action: v3.0.0 → v3.4.0
* docker/metadata-action: v5.5.1 → v5.7.0
* docker/build-push-action: v5.1.0 → v6.17.0

The other change is to start using commit hashes instead of tags, when
specifying the specific version of a GitHub Action.  This change is in
response to the tj-actions/changed-files supply-chain attach: The
attackers were able to change which commit an existing tag pointed to,
so all the GitHub Actions workflows pointing to the tag ran the
attacker's code.

The partial protection to this is to specify commit IDs, since commit
IDs are immutable.  It's not full protection, in case the Actions we
refer to are themselves referring to things by tag name.

In the future, Dependabot will create PRs to update GitHub Actions.
Dependabot is smart enough to recognize that we're using commit hashes,
and will submit PRs that use hashes: It will automatically convert the
Action's new-version tag into a commit hash.