Workflow: Mirror main branch to MGP-Public repo
The purpose of this workflow is to keep the MGP-Public repo's main branch in
sync this this repo's main branch.  This new workflow should run any time there
is a push to the main branch.  Only one instance of this workflow should be
running at once: This is unlikely to happen, but could happen if many pushes to
main are happening in a short time period.

The workflow works by checking out both this main branch and MGP-Public's main
branch.  The workflow then descends into the MGP-Public checkout, adds MGP's
checkout as a remote (a local remote), and then does a fast-forward merge from
MGP main to MGP-Public main.

After the above merge happens, a shortlog of changes is printed, and then the
newly-merged MGP-Public main branch is pushed to GitHub.

Since MGP-Public's main branch is meant to be identical to MGP main, we need to
make sure GitHub does not try to run this workflow in the MGP-Public repo.  To
prevent this from happening, the mirror job is conditioned to run only if the
"DO_MIRROR" repository variable exists and is set to "yes".  In the MGP-Public
repo, DO_MIRROR either does not exist, or is set to some other value (like
"no").

Here is the mapping of GitHub Actions version to commit hash:

actions/checkout@v5.0.0 -> 08c6903cd8c0fde910a37f88322edcfb5dd907a8