README: Add information about Demo Mode
MGP now supports Demo Mode: When MGP is run with default database
configuration (that is, the database connection information encoded in
the Dockerfile), MGP will not attempt to connect to a mTilda database.
Instead, it will use a built-in SQLite3 database, containing two
'mTilda' users and three patients.

This commit makes a number of additions:

* A teaser is added near the top of the README.

* The *ODBC* section includes a note about how demo mode does not require
  ODBC, but instead uses RSQLite with a built-in database.

* After the *Reference Databases* section, a new section is added:
  *Demonstration Data* points out the location of the demonstration data
  SQL dump, and what's included in it.  This section will likely be
  expanded later.

* The *How to Build* section sees a large amount of content added under
  *Building from Source*: Not only does the user need to build the
  SQLite3 database, enabling demo mode requires setting environment
  variables to specific case-sensitive values.

* Finally, the *How to Run* section has a new sub-section added, talking
  specifically about demo mode.  This is where the user can find the
  'mTilda' usernames and patient ITLs needed to demonstrate MGP.

There will be at least on subsequent commit, providing additional
information about the demonstration data.