README: Rewrite the build instructions
This is a large rewrite of the "How to Build" instructions of the
README, with the idea of making it more approachable for less
technically-savvy users.

First of all, the section is now called "How to Get", because the first
method of building is to not build at all, but instead is to get the
pre-built container image from GitHub.  The first two methods have also
had their names changed to be more descriptive.

NOTE: The instructions intentionally point to the public repo here, not
the internal repo.

In the intro of the section, text has been added to point users towards
Docker Desktop, along with a pointer to the relevent sections of the
Docker Desktop documentation.  I include a note about MGP being meant to
be run as a web application, not on an individual's machine.  And I add
a note that, if the end user is going to be doing this an a work-owned
machine, they should be talking to their IT support folks first.

The "Ready-Made Container Image" method now puts the Docker command up
top, followed with an explanation of what is happening, and finishing
with a description of what they should see in Docker Desktop, assuming
the pull was successful.

The next method removes the need for a Git client, instead pointing the
user to download a ZIP file of the archive from GitHub.  I note that
this method requires intermediate skills.  Again, the explanation of
what is happening is moved to be after the command that needs to be run.

The "Build from Source" method is not changed, other than adding a note
that advanced IT and R skills are required.