README: Clean up environment variable docs
This commit rewrites and moves around the documentation related to environment
variables.

The "Configuring the ODBC Driver" section is moved from near-top of the "How to
Run" section, and now lives in the "Running from Source" section.  This is
appropriate because the only time you need these instructions is when you're
running a built-from-source MGP.  As a result of the migration, some of the
"Running from Source" text is moved around.

The rest of the "Environment Variables" section is also cleaned up.  Readers
are now told to assume that each environment variable is required, and the
individual environment variables will tell you when it is either optional, or
should not be set (typically because of Demo Mode).

The entry for the `DRIVER` environment variable is moved to the end of the
list, since it only need to be set in one case.