README: Add information on how to run demo mode
This is a major rewrite of the start of the "How to Run" section, along with
the demo mode sub-section.  The content now flows directly from the start of
"How to Run", into the demo mode sub-section, and it takes a form similar to
the "How to Build" sections.  A couple of screenshots are also added.